third_party/WebKit/LayoutTests/fast/css/namespaces/empty-namespace-compound-selector-1.xhtml

<html xmlns="http://www.w3.org/1999/xhtml">
<style>
@namespace url(foo);

*|div {
    display: block;
    width: 100px;
    height: 100px;
    background-color: red;
}

|div#target {
    background-color: green;
}
</style>
<div id="target" xmlns=""></div>

<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>

<script>
test(function(){
    assert_equals(getComputedStyle(document.getElementById('target')).backgroundColor, 'rgb(0, 128, 0)');
}, "Empty namespaces should work correctly on compound selectors");
</script>

</html>
